+++ src/alley/wp/alleyvate/features/class-user-enumeration-restrictions.php
@@ -10,7 +10,7 @@
  * @package wp-alleyvate
  */
 
-namespace Alley\WP\Alleyvate\Feature;
+namespace Alley\WP\Alleyvate\Features;
 
 use Alley\WP\Alleyvate\Feature;
 use WP_Error;
